move utils to separate files within utils dir
