+++ src/plugins/SimpleRule.js
