+++ src/utils/findRenderer.js
@@ -1,6 +1,6 @@
 import isInBrowser from 'is-in-browser'
-import DomRenderer from './backends/DomRenderer'
-import VirtualRenderer from './backends/VirtualRenderer'
+import DomRenderer from '../backends/DomRenderer'
+import VirtualRenderer from '../backends/VirtualRenderer'
 
 /**
  * Find proper renderer.
